Factor out common elements, tidy up tests, spelling
